feat: GroupList -> Groups로 이름 변경하고 GroupList에는 실제 Group list만 보이도록 분리
