+++ src/components/organisms/group/index.tsx
@@ -1,19 +1,13 @@
 'use client';
 
 import { Filter } from '@/components/molecules/group/filter';
-import { GroupCard } from '@/components/molecules/group/group-card';
+import { GroupList } from '@/components/molecules/group/group-list';
 import { SortOrder } from '@/components/molecules/group/sort-order';
 import { SearchInput } from '@/components/molecules/search-input/search-input';
 import { Tab, TabType } from '@/components/molecules/tab';
-import { getBookmarkList } from '@/features/bookmark';
-import { useFetchInView } from '@/hooks/useFetchInView';
-import { useFetchItems } from '@/hooks/useFetchItems';
-import useAuthStore from '@/stores/useAuthStore';
-import { Group, GroupType } from '@/types';
-import { Position, Skill } from '@/types/enums';
-import flattenPages from '@/utils/flattenPages';
+import { GroupType } from '@/types';
 import { useRouter } from 'next/navigation';
-import { useEffect, useMemo, useState } from 'react';
+import { Suspense } from 'react';
 
 type GroupListProps = {
   searchParams: Record<string, string | undefined>;
@@ -25,20 +19,10 @@ const tabList: TabType[] = [
   { value: GroupType.PROJECT, label: '프로젝트' },
 ];
 
-enum EMPTY_INFO_MESSAGE {
-  EMPTY_INITIAL = '생성된 그룹이 없습니다',
-  SEARCH = '검색 결과가 없습니다',
-  FILTER = '조건에 해당하는 그룹이 없습니다.',
-}
-
-export const GroupList = ({ searchParams }: GroupListProps) => {
-  const [isEmptyItems, setIsEmptyItems] = useState(true);
-  const [emptyInfoMessage, setEmptyInfoMessage] =
-    useState<EMPTY_INFO_MESSAGE | null>(null);
+export const Groups = ({ searchParams }: GroupListProps) => {
+  console.log('client', searchParams);
   const router = useRouter();
-  const user = useAuthStore((state) => state.user);
-
-  /**
+  /*
    * router.push를 수행하는 함수
    * @param queries 여러 query key를 한번에 업데이트 할 수 있기 때문에 인자를 Record 타입으로 받는다
    */
@@ -69,77 +53,6 @@ export const GroupList = ({ searchParams }: GroupListProps) => {
     router.push(`?${params.toString()}`);
   };
 
-  const queryParams = useMemo(
-    () => ({
-      type: searchParams.type ?? '',
-      skills: Skill[searchParams.skill as keyof typeof Skill] ?? '',
-      position: Position[searchParams.position as keyof typeof Position] ?? '',
-      sort: searchParams.sort ?? 'createdAt',
-      order: searchParams.order ?? 'desc',
-      cursor: searchParams.order === 'desc' || !searchParams.order ? 'null' : 0,
-      search: searchParams.search ?? '',
-    }),
-    [searchParams],
-  );
-
-  const { data, fetchNextPage, hasNextPage, isLoading } = useFetchItems<Group>({
-    url: '/v2/groups',
-    queryParams: { ...queryParams, size: 10 },
-  });
-
-  const items = flattenPages(data.pages);
-
-  const { ref } = useFetchInView({
-    fetchNextPage,
-    isLoading,
-    options: {
-      rootMargin: '50px',
-    },
-  });
-
-  useEffect(() => {
-    if (items.length === 0) {
-      // 받아온 데이터가 없는 경우
-      setIsEmptyItems(true);
-      if (searchParams.search) {
-        // 검색어가 있다면 검색어를 우선으로 메시지 설정
-        setEmptyInfoMessage(EMPTY_INFO_MESSAGE.SEARCH);
-        return;
-      } else if (
-        searchParams.type ||
-        searchParams.skill ||
-        searchParams.position
-      ) {
-        setEmptyInfoMessage(EMPTY_INFO_MESSAGE.FILTER);
-        return;
-      }
-      setEmptyInfoMessage(EMPTY_INFO_MESSAGE.EMPTY_INITIAL); // 받아온 데이터는 없지만 필터도 없는 경우(아직 생성된 그룹이 하나도 없을 경우)
-      return;
-    }
-    setEmptyInfoMessage(null);
-    setIsEmptyItems(false);
-  }, [searchParams, items.length]);
-
-  // useEffect(() => {
-  //   console.log('✅ Hydrated data from client:', queryParams); // DEV : 💡 서버 컴포넌트에서 prefetch 하는지 확인용
-  // }, [queryParams]);
-
-  //북마크 처리
-  const [displayItems, setDisplayItems] = useState<Group[]>(items);
-
-  useEffect(() => {
-    if (!user) {
-      const bookmark = getBookmarkList();
-      const processedItems = flattenPages(data.pages).map((item) => ({
-        ...item,
-        isBookmark: bookmark.includes(item.id),
-      }));
-      setDisplayItems(processedItems);
-    } else {
-      setDisplayItems(flattenPages(data.pages));
-    }
-  }, [data.pages, user]);
-
   return (
     <>
       <Tab
@@ -151,16 +64,9 @@ export const GroupList = ({ searchParams }: GroupListProps) => {
           <SortOrder updateQueryParams={updateQueryParams} />
         </div>
         <SearchInput />
-        {isEmptyItems && emptyInfoMessage !== null ? (
-          <div>{emptyInfoMessage}</div>
-        ) : (
-          <ul className="flex flex-col gap-3 mt-8 md:flex-row md:flex-wrap md:gap-6 md:justify-center">
-            {displayItems.map((group) => (
-              <GroupCard key={group.id} item={group} />
-            ))}
-          </ul>
-        )}
-        {hasNextPage && <div ref={ref}></div>}
+        <Suspense fallback={<div>Loading...</div>}>
+          <GroupList searchParams={searchParams} />
+        </Suspense>
       </Tab>
     </>
   );
